+++ allinone/ns-allinone-2.35/ns-2.35/pause/pause_header.h
@@ -8,7 +8,7 @@
    static const int ETH_CTRL_FRAME_SIZE = 64;
    static const size_t NUM_ETH_CLASS = 8;
  
-   uint16_t class_pause_durations_[NUM_ETH_CLASS];
+   double class_pause_durations_[NUM_ETH_CLASS];
    bool class_enable_vector_[NUM_ETH_CLASS];
  
    /* Packet header access functions */
@@ -23,7 +23,7 @@
  
    /* Fill in fields */
    static void fill_in(Packet* p,
-                const std::vector<uint16_t> & s_class_pause_durations,
+                const std::vector<double> & s_class_pause_durations,
                 const std::vector<bool> & s_class_enable_vector);
  };
